migrate cloud unit tests to pytest
